refactor: move contract utils to another crate
