+++ crates/l2/sdk/contract_utils/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "ethrex-sdk-contract-utils"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+thiserror.workspace = true
+tracing.workspace = true
