+++ crates/l2/sdk/src/sdk.rs
@@ -1,5 +1,3 @@
-use std::path::PathBuf;
-use std::process::{Command, ExitStatus};
 use std::{fs::read_to_string, path::Path};
 
 use bytes::Bytes;
@@ -18,10 +16,13 @@ use secp256k1::SecretKey;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 pub mod calldata;
-pub mod l1_to_l2_tx_data;
+mod l1_to_l2_tx_data;
 
 pub use l1_to_l2_tx_data::{L1ToL2TransactionData, send_l1_to_l2_tx};
-use tracing::{info, trace};
+
+// Reexport the contracts module
+#[doc(inline)]
+pub use ethrex_sdk_contract_utils::*;
 
 // 0x8ccf74999c496e4d27a2b02941673f41dd0dab2a
 pub const DEFAULT_BRIDGE_ADDRESS: Address = H160([
@@ -345,91 +346,6 @@ pub fn get_address_from_secret_key(secret_key: &SecretKey) -> Result<Address, Et
     Ok(Address::from(address_bytes))
 }
 
-#[derive(Debug, thiserror::Error)]
-pub enum ContractCompilationError {
-    #[error("The path is not a valid utf-8 string")]
-    FailedToGetStringFromPath,
-    #[error("Deployer compilation error: {0}")]
-    CompilationError(String),
-    #[error("Could not read file")]
-    FailedToReadFile(#[from] std::io::Error),
-    #[error("Failed to serialize/deserialize")]
-    SerializationError(#[from] serde_json::Error),
-}
-
-pub fn compile_contract(
-    general_contracts_path: &Path,
-    contract_path: &str,
-    runtime_bin: bool,
-) -> Result<(), ContractCompilationError> {
-    let bin_flag = if runtime_bin {
-        "--bin-runtime"
-    } else {
-        "--bin"
-    };
-
-    // Both the contract path and the output path are relative to where the Makefile is.
-    if !Command::new("solc")
-        .arg(bin_flag)
-        .arg(
-            "@openzeppelin/contracts=".to_string()
-                + general_contracts_path
-                    .join("lib")
-                    .join("openzeppelin-contracts-upgradeable")
-                    .join("lib")
-                    .join("openzeppelin-contracts")
-                    .join("contracts")
-                    .to_str()
-                    .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
-        )
-        .arg(
-            "@openzeppelin/contracts-upgradeable=".to_string()
-                + general_contracts_path
-                    .join("lib")
-                    .join("openzeppelin-contracts-upgradeable")
-                    .join("contracts")
-                    .to_str()
-                    .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
-        )
-        .arg(
-            general_contracts_path
-                .join(contract_path)
-                .to_str()
-                .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
-        )
-        .arg("--via-ir")
-        .arg("-o")
-        .arg(
-            general_contracts_path
-                .join("solc_out")
-                .to_str()
-                .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
-        )
-        .arg("--overwrite")
-        .arg("--allow-paths")
-        .arg(
-            general_contracts_path
-                .to_str()
-                .ok_or(ContractCompilationError::FailedToGetStringFromPath)?,
-        )
-        .spawn()
-        .map_err(|err| {
-            ContractCompilationError::CompilationError(format!("Failed to spawn solc: {err}"))
-        })?
-        .wait()
-        .map_err(|err| {
-            ContractCompilationError::CompilationError(format!("Failed to wait for solc: {err}"))
-        })?
-        .success()
-    {
-        return Err(ContractCompilationError::CompilationError(
-            format!("Failed to compile {contract_path}").to_owned(),
-        ));
-    }
-
-    Ok(())
-}
-
 // 0x4e59b44847b379578588920cA78FbF26c0B4956C
 const DETERMINISTIC_CREATE2_ADDRESS: Address = H160([
     0x4e, 0x59, 0xb4, 0x48, 0x47, 0xb3, 0x79, 0x57, 0x85, 0x88, 0x92, 0x0c, 0xa7, 0x8f, 0xbf, 0x26,
@@ -650,195 +566,4 @@ pub async fn call_contract(
 
     wait_for_transaction_receipt(tx_hash, client, 100).await?;
     Ok(tx_hash)
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum GitError {
-    #[error("Failed to clone: {0}")]
-    DependencyError(String),
-    #[error("Internal error: {0}")]
-    InternalError(String),
-    #[error("Failed to get string from path")]
-    FailedToGetStringFromPath,
-}
-
-pub fn git_clone(
-    repository_url: &str,
-    outdir: &str,
-    branch: Option<&str>,
-    submodules: bool,
-) -> Result<ExitStatus, GitError> {
-    info!(repository_url = %repository_url, outdir = %outdir, branch = ?branch, "Cloning or updating git repository");
-
-    if PathBuf::from(outdir).join(".git").exists() {
-        info!(outdir = %outdir, "Found existing git repository, updating...");
-
-        let branch_name = if let Some(b) = branch {
-            b.to_string()
-        } else {
-            // Look for default branch name (could be main, master or other)
-            let output = Command::new("git")
-                .current_dir(outdir)
-                .arg("symbolic-ref")
-                .arg("refs/remotes/origin/HEAD")
-                .output()
-                .map_err(|e| {
-                    GitError::DependencyError(format!(
-                        "Failed to get default branch for {outdir}: {e}"
-                    ))
-                })?;
-
-            if !output.status.success() {
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                return Err(GitError::DependencyError(format!(
-                    "Failed to get default branch for {outdir}: {stderr}"
-                )));
-            }
-
-            String::from_utf8(output.stdout)
-                .map_err(|_| GitError::InternalError("Failed to parse git output".to_string()))?
-                .trim()
-                .split('/')
-                .next_back()
-                .ok_or(GitError::InternalError(
-                    "Failed to parse default branch".to_string(),
-                ))?
-                .to_string()
-        };
-
-        trace!(branch = %branch_name, "Updating to branch");
-
-        // Fetch
-        let fetch_status = Command::new("git")
-            .current_dir(outdir)
-            .args(["fetch", "origin"])
-            .spawn()
-            .map_err(|err| GitError::DependencyError(format!("Failed to spawn git fetch: {err}")))?
-            .wait()
-            .map_err(|err| {
-                GitError::DependencyError(format!("Failed to wait for git fetch: {err}"))
-            })?;
-        if !fetch_status.success() {
-            return Err(GitError::DependencyError(format!(
-                "git fetch failed for {outdir}"
-            )));
-        }
-
-        // Checkout to branch
-        let checkout_status = Command::new("git")
-            .current_dir(outdir)
-            .arg("checkout")
-            .arg(&branch_name)
-            .spawn()
-            .map_err(|err| {
-                GitError::DependencyError(format!("Failed to spawn git checkout: {err}"))
-            })?
-            .wait()
-            .map_err(|err| {
-                GitError::DependencyError(format!("Failed to wait for git checkout: {err}"))
-            })?;
-        if !checkout_status.success() {
-            return Err(GitError::DependencyError(format!(
-                "git checkout of branch {branch_name} failed for {outdir}, try deleting the repo folder"
-            )));
-        }
-
-        // Reset branch to origin
-        let reset_status = Command::new("git")
-            .current_dir(outdir)
-            .arg("reset")
-            .arg("--hard")
-            .arg(format!("origin/{branch_name}"))
-            .spawn()
-            .map_err(|err| GitError::DependencyError(format!("Failed to spawn git reset: {err}")))?
-            .wait()
-            .map_err(|err| {
-                GitError::DependencyError(format!("Failed to wait for git reset: {err}"))
-            })?;
-
-        if !reset_status.success() {
-            return Err(GitError::DependencyError(format!(
-                "git reset failed for {outdir}"
-            )));
-        }
-
-        // Update submodules
-        if submodules {
-            let submodule_status = Command::new("git")
-                .current_dir(outdir)
-                .arg("submodule")
-                .arg("update")
-                .arg("--init")
-                .arg("--recursive")
-                .spawn()
-                .map_err(|err| {
-                    GitError::DependencyError(format!(
-                        "Failed to spawn git submodule update: {err}"
-                    ))
-                })?
-                .wait()
-                .map_err(|err| {
-                    GitError::DependencyError(format!(
-                        "Failed to wait for git submodule update: {err}"
-                    ))
-                })?;
-            if !submodule_status.success() {
-                return Err(GitError::DependencyError(format!(
-                    "git submodule update failed for {outdir}"
-                )));
-            }
-        }
-
-        Ok(reset_status)
-    } else {
-        trace!(repository_url = %repository_url, outdir = %outdir, branch = ?branch, "Cloning git repository");
-        let mut git_cmd = Command::new("git");
-
-        let git_clone_cmd = git_cmd.arg("clone").arg(repository_url);
-
-        if let Some(branch) = branch {
-            git_clone_cmd.arg("--branch").arg(branch);
-        }
-
-        if submodules {
-            git_clone_cmd.arg("--recurse-submodules");
-        }
-
-        git_clone_cmd
-            .arg(outdir)
-            .spawn()
-            .map_err(|err| GitError::DependencyError(format!("Failed to spawn git: {err}")))?
-            .wait()
-            .map_err(|err| GitError::DependencyError(format!("Failed to wait for git: {err}")))
-    }
-}
-
-pub fn download_contract_deps(contracts_path: &Path) -> Result<(), GitError> {
-    trace!("Downloading contract dependencies");
-    std::fs::create_dir_all(contracts_path.join("lib")).map_err(|err| {
-        GitError::DependencyError(format!("Failed to create contracts/lib: {err}"))
-    })?;
-
-    git_clone(
-        "https://github.com/OpenZeppelin/openzeppelin-contracts-upgradeable.git",
-        contracts_path
-            .join("lib/openzeppelin-contracts-upgradeable")
-            .to_str()
-            .ok_or(GitError::FailedToGetStringFromPath)?,
-        None,
-        true,
-    )?;
-
-    git_clone(
-        "https://github.com/succinctlabs/sp1-contracts.git",
-        contracts_path
-            .join("lib/sp1-contracts")
-            .to_str()
-            .ok_or(GitError::FailedToGetStringFromPath)?,
-        None,
-        false,
-    )?;
-
-    trace!("Contract dependencies downloaded");
-    Ok(())
 }
